Initial portieris-armada-app deliverable
This commit introduces the stx-portieris armada app, which is built
through build-helm-charts.sh, which is to be manually uploaded and
applied.

In order to build this app, the portieris-helm and stx-portieris-helm
packages are introduced, with portieris-helm repackaging the helm
charts from upstream portieris, and stx-portieris-helm packaging these
charts to be built into an armada app.

These packages are not included as part of an iso install, so will have
no affect unless the armada application is uploaded.

Placeholder certificates are currently included, and should be removed
once cert-manager integration is complete.  This is unsuitable for a
real-world deployment, and should not be used in production, due to
the availability of the private key to an attacker.

Change-Id: I806063c0f561a20808b4cbc4c4a561be5c806c5b
Story: 2007348
Task: 39597
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>