add toggle to reinstall the admission webhook
The backup and restore procedure will set ReapplyAdmissionWebhook to
true to force an "upgrade" of an already installed Portieris
application.  The webhooks job is run on the helm chart's post-upgrade
hook.  The chart value ReapplyAdmissionWebhook is reserved to simplify
its handling by backup and restore.

This resolves the issue of forcing the webhook to reapply after it has
been manually removed before a backup operation.  The 'post-install'
helm.sh/hook for example is only run after the chart is installed.  We
need to trigger the webhook creation after backup and restore
operations.  Setting an application override as follows will trigger an
upgrade of the helm chart when the application is reapplied:

  system helm-override-update \
    --reuse-values \
    portieris portieris portieris \
    --set ReapplyAdmissionWebhook=true

Partial-Bug: 1890630
Change-Id: I4682765efddc217e792b37c659ae5833379bf054
Signed-off-by: Michel Thebeau <Michel.Thebeau@windriver.com>