Upversion Portieris to 0.13.10
This commit upversions Portieris to version 0.13.10. The upversioning
addresses CVEs in the previons (0.13.1) Portieris image.

The additional toleration is to allow Portieris pods to be scheduled.
Without it, application apply fails. The taint used to be "master"
in older releases, but has been changed to "control-plane". Keeping
both for backwards compatibility and upgrades.

Test Cases:

PASS: Apply new verison of Portieris. Apply an imagepolicy. Ensure
      that a signed image can be used and an unsigned image is denied
      according to the imagepolicy.
PASS: Remove all user created imagepolicies and try to use the
      unsigned image. Ensure that by default, unsigned images are
      allowed.

Closes-bug: 2051611
Change-Id: Id621d91ed41a705035713ff59439f59211e035f5
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>